Refactor ProfileActivity into separate extension files
Extracts avatar handling, birth date logic, remote profile operations, and related models from ProfileActivity. This improves modularity and readability by separating concerns into dedicated extension files.
